Se agregan servicios al controlador persona para separar la logica de negocio, se añaden mensajes personalizados de excepciones

+++ src/main/java/com/inventario/parque/kyqa/InventarioKyqa/Repository/InventarioRepository.java
@@ -4,5 +4,5 @@ import com.inventario.parque.kyqa.InventarioKyqa.Entity.Inventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface InventarioRepository extends JpaRepository<Inventario, Long>, JpaSpecificationExecutor<Inventario> {
+public interface InventarioRepository extends JpaRepository<Inventario, Integer>, JpaSpecificationExecutor<Inventario> {
 }
